easy_r1/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "verl"
dependencies = [
    "accelerate",
    "alphashape>=1.3.1",
    "codetiming",
    "datasets",
    "flash-attn>=2.4.3",
    "joblib>=1.3.0",
    "liger-kernel",
    "mathruler",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "omegaconf",
    "opencv-python>=4.8.0",
    "openai>=1.0.0",
    "pandas",
    "peft",
    "pillow>=10.0.0",
    "plyfile>=1.0.0",
    "pyarrow>=15.0.0",
    "pylatexenc",
    "pyyaml>=6.0",
    "qwen-vl-utils",
    "ray[default]",
    "scipy>=1.10.0",
    "shapely>=2.0.0",
    "tensordict",
    "torchdata",
    "tqdm>=4.66.0",
    "transformers>=4.54.0,<=4.57.0",
    "vllm>=0.8.0",
    "wandb",
]
dynamic = [
    "version",
    "requires-python",
    "authors",
    "description",
    "readme",
    "license"
]

[project.optional-dependencies]
dev = ["pre-commit", "ruff"]

[tool.ruff]
target-version = "py39"
line-length = 119
indent-width = 4

[tool.ruff.lint]
ignore = ["C901", "E501", "E741", "W605", "C408"]
select = ["C", "E", "F", "I", "W", "RUF022"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F811"]

[tool.ruff.lint.isort]
lines-after-imports = 2
known-first-party = ["verl"]
known-third-party = ["torch", "transformers", "wandb"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
